REQ-0007: Convert PowerMonNodes to Linux from Windows
Converting files from Windows OS to Linux

This commit has all threads running correctly, and has a working set of
sockets for talking to the PowerMon server.

This commit also has incorporated the Avahi service thread functionality
and does see the server and is able to get configuration parameters.

However, the singletons don't appear to be working correctly. I cannot
put messages on the wifi queue while I can put them on the screen using
the console, which is a vritual imitation of the wifi implementation.
So, following this commit I am going to move away from singletons to
message queues.